requirements.txt
flask
Flask-Alembic
flask-bcrypt
flask-login
flask-migrate
flask-sqlalchemy-lite
psycopg2
pytest
pytest-cov
python-dotenv
requests
